rename RenderMapProps file
