+++ packages/ui-components/src/components/Maps/USNationalMap/interfaces.ts
@@ -1,4 +1,4 @@
-import { RenderMapProps } from "../interfaces";
+import { RenderMapProps } from "../RenderMapProps";
 import { Metric } from "@actnowcoalition/metrics";
 
 export interface USNationalMapProps extends RenderMapProps {
